StreamLit/requirements.txt
streamlit
pillow
torch
torchvision
